shared/ui/index.tsx 파일 내 셀렉트 컴포넌트 분리
