+++ src/shared/ui/card/index.ts
@@ -1,2 +1,2 @@
-export { Card, CardHeader, CardTitle, CardContent } from "./card"
+export { Card, CardHeader, CardTitle, CardContent } from "./Card"
 export type { CardProps, CardHeaderProps, CardTitleProps, CardContentProps } from "./types"
