+++ src/shared/ui/select/index.ts
@@ -1,0 +1,3 @@
+export { Select, SelectGroup, SelectValue, SelectTrigger, SelectContent, SelectItem } from "./Select"
+
+export type { SelectTriggerProps, SelectContentProps, SelectItemProps } from "./types"
